allow running workspace in Ram disk
By default, a `emptyDir` volume on Kubernetes is hosted by the
underlying node disk. If this disk is not an SSD, this can have
a big impact on the performance of Jenkins jobs.

This allows you to load the entire workspace into memory,
providing a huge speed improvement, at the cost of
needing more memory for each job.